修改README

+++ lib/net.h
@@ -26,7 +26,7 @@ typedef struct
 } g_socket;
 
 g_socket *tcp_socket_client(const char* address, u_int16_t port);
-g_socket *tcp_socket_server(u_int16_t port);
+g_socket *tcp_socket_server(unsigned port);
 void close_socket(g_socket *socket);
 int tcp_receive(g_socket *sock, char* buffer, unsigned buf_size, int timeout);
 void *get_localip(char *ip);
